requirements.txt
beautifulsoup4==4.12.3
feedparser==6.0.11
numpy==2.1.2
pandas==2.2.3
pillow==10.4.0
requests==2.32.3
sentence-transformers==3.0.1
sentencepiece==0.2.0
streamlit==1.37.1
torch==2.4.1
tqdm==4.66.5
transformers==4.44.2
